link positions into separate file
