--- apps/markdownlinks/format.js
@@ -1,165 +1,0 @@
-const format = (text) => {
-  /* receive markdown hyperlink syntax, return slack hyperlink syntax */
-  var bracketsParentheses = allIndexOf(text, '](')
-  var brackets = allIndexOf(text, '[')
-  var parentheses = allIndexOf(text, ')')
-  if (bracketsParentheses) {
-    let message = text
-    var foundAllLinkPositions = allLinkPositions(bracketsParentheses, brackets, parentheses)
-    for (var linkPositions of foundAllLinkPositions) {
-      if (validLinkPositions(linkPositions)) {
-        var linkString = findLinkString(linkPositions, text)
-        var linkAddress = findLinkAddress(linkPositions, text)
-        var linkAddressTrimmed = linkAddress.trim()
-        var linkAddressTrimmedLowered = linkAddressTrimmed.toLowerCase()
-        if (checkLinkString(linkString) && checkLinkAddress(linkAddressTrimmedLowered)) {
-          var linkAddressHttped = httpLinkAddress(linkAddressTrimmedLowered)
-          var messageLink = createMessageLink(linkAddressHttped, linkString)
-          var markdownLink = findMarkdownLink(linkPositions, text)
-          message = replaceLink(markdownLink, messageLink, message)
-        }
-      }
-    }
-    return message
-  } else {
-    return text
-  }
-}
-
-// Functions to check for threats
-
-const allIndexOf = (text, searchChar) => {
-  /* find all the positions of a character in a string */
-  let startIndex = 0; let index; const indices = []; let count = 0
-  while ((index = text.indexOf(searchChar, startIndex)) > -1 && count < 20) {
-    indices.push(index)
-    startIndex = index + 1
-    count++
-  }
-  return indices
-}
-
-const allLinkPositions = (bracketsParentheses, brackets, parentheses) => {
-  /* all of the positions of characters which compose markdown syntax links.
-  a closed bracket/open parenthesis pair is used as the link indicator */
-  var allPositions = []
-  var bracketsParenthesesLen = bracketsParentheses.length
-  for (var i = 0; i < bracketsParenthesesLen; i++) {
-    var previousPosition = findPreviousPosition(i, bracketsParentheses)
-    var currentPosition = bracketsParentheses[i]
-    var nextPosition = findNextPosition(i, bracketsParentheses, bracketsParenthesesLen, parentheses)
-
-    var positions = [undefined, currentPosition, undefined]
-    positions[0] = findOpenBracket(brackets, currentPosition, previousPosition)
-    positions[2] = findClosedParenthensis(parentheses, currentPosition, nextPosition)
-
-    allPositions.push(positions)
-  }
-  return allPositions
-}
-
-const findPreviousPosition = (i, bracketsParentheses) => {
-  /* find the bracket/parenthesis pair that occurs before the current one */
-  if (i === 0) {
-    return 0
-  } else {
-    var j = i - 1
-    return bracketsParentheses[j]
-  }
-}
-
-const findNextPosition = (i, bracketsParentheses, bracketParenthesesLen, parentheses) => {
-  /* find the bracket/parenthesis pair that occurs after the current one */
-  if (i === (bracketParenthesesLen - 1)) {
-    return parentheses[parentheses.length - 1]
-  } else {
-    var k = i + 1
-    return bracketsParentheses[k]
-  }
-}
-
-const findClosedParenthensis = (parentheses, currentPosition, nextPosition) => {
-  /* find the position of the closed parenthesis, associated with the hyperlink */
-  var filteredParentheses = parentheses.filter(parenthesis => parenthesis > currentPosition && parenthesis <= nextPosition)
-  return filteredParentheses[0]
-}
-
-const findOpenBracket = (brackets, currentPosition, previousPosition) => {
-  /* find the position of the open bracket associated with the hyperlink */
-  var filteredBrackets = brackets.filter(bracket => bracket < currentPosition && bracket >= previousPosition)
-  return filteredBrackets.pop()
-}
-
-const validLinkPositions = (linkPositions) => {
-  /* check that the set of positions for characters could represent a hyperlink */
-  var hasValues = linkPositions.every(value => value >= 0)
-  var hasNumbers = linkPositions.every(value => typeof (value) === 'number')
-  var correctLength = linkPositions.length === 3
-  var correctOrder = (linkPositions[0] < linkPositions[1] && linkPositions[1] < linkPositions[2])
-  return (correctLength && hasValues && hasNumbers && correctOrder)
-}
-
-const findMarkdownLink = (linkPositions, text) => {
-  /* identify entire portion of markdown syntax */
-  return text.slice(linkPositions[0], linkPositions[2] + 1)
-}
-
-const findLinkString = (linkPositions, text) => {
-  /* identify text portion of hyperlink from message string */
-  return text.slice(linkPositions[0] + 1, linkPositions[1])
-}
-
-const checkLinkString = (linkString) => {
-  /* link string cannot be blank or only spaces */
-  var linkStringTrim = linkString.trim()
-  if (linkStringTrim) {
-    return true
-  } else {
-    return false
-  }
-}
-
-const findLinkAddress = (linkPositions, text) => {
-  /* identify url portion of link from message string */
-  return text.slice(linkPositions[1] + 2, linkPositions[2])
-}
-
-const checkLinkAddress = (linkAddress) => {
-  /* must fit the correct format of a url */
-  var domainRegex = /^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([-.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/
-  if (linkAddress.match(domainRegex)) {
-    return true
-  } else {
-    return false
-  }
-}
-
-const httpLinkAddress = (linkAddress) => {
-  /* ensure that each link has http or https in the url */
-  if (linkAddress.startsWith('http://') || linkAddress.startsWith('https://')) {
-    return linkAddress
-  } else {
-    return `https://${linkAddress}`
-  }
-}
-
-const createMessageLink = (linkAddress, displayText) => {
-  /* create slack syntax for text and url */
-  return `<${linkAddress}|${displayText}>`
-}
-
-const replaceLink = (markdownLink, messageLink, message) => {
-  /* identify and replace the hyperlink based on its exact structure */
-  return message.replace(markdownLink, messageLink, message)
-}
-
-module.exports = {
-  format,
-  allIndexOf,
-  allLinkPositions,
-  validLinkPositions,
-  findMarkdownLink,
-  httpLinkAddress,
-  checkLinkString,
-  checkLinkAddress
-}
