+++ apps/markdownlinks/format/format.test.js
@@ -1,6 +1,5 @@
 const {
-  format, allIndexOf, allLinkPositions,
-  validLinkPositions, findMarkdownLink, httpLinkAddress,
+  format, findMarkdownLink, httpLinkAddress,
   checkLinkString, checkLinkAddress
 } = require('./format')
 
@@ -29,32 +28,6 @@ var formatText = [
 test.each(formatText)(
   'receive markdown hyperlink syntax, return slack hyperlink syntax', (testMessage, expectedMessage) => {
     expect(format(testMessage)).toEqual(expectedMessage)
-  })
-
-/* base arrays for character position tests */
-var bracketsParentheses = [16, 52, 91]
-var parentheses = [28, 31, 75, 104]
-var brackets = [4, 12, 38, 78]
-
-var charPositions = [
-  [testMessageOne, '](', bracketsParentheses],
-  [testMessageOne, ')', parentheses],
-  [testMessageOne, '[', brackets]]
-test.each(charPositions)(
-  'finds all the positions of a character in a string', (text, char, expectedArray) => {
-    expect(allIndexOf(text, char)).toEqual(expectedArray)
-  })
-
-var foundAllLinkPositions = [[12, 16, 28], [38, 52, 75], [78, 91, 104]]
-test.each([[bracketsParentheses, brackets, parentheses, foundAllLinkPositions]])(
-  'all of the positions of characters which compose markdown syntax links', (bracketsParentheses, brackets, parentheses, expectedArray) => {
-    expect(allLinkPositions(bracketsParentheses, brackets, parentheses)).toEqual(expectedArray)
-  })
-
-var isValidLinkPositions = [[[12, 16, 28], true], [[undefined, 34, 23], false], [[56], false], [[12, 6, 20], false], [['0', 1, 2], false]]
-test.each(isValidLinkPositions)(
-  'check that the set of positions for characters could represent a hyperlink', (linkPositions, expectedBoolean) => {
-    expect(validLinkPositions(linkPositions)).toBe(expectedBoolean)
   })
 
 var markdownLink = [[[12, 16, 28], testMessageOne, '[car](dmv.ca.gov)'], [[38, 52, 75], testMessageOne, '[safest of all](https://www.osha.com/)'], [[78, 91, 104], testMessageOne, '[Example site](example.com)']]
